Added a sorting feature on the homepage
Added a select HTML element for selecting a sorting option. When one
option is selected, a submit action takes place immediately (onChange -submit) and a querystring is sent to the backend;
A new queryset is produced according to querystring (as accessed by its associated parameter in the request dict);

Issues arise when combining multiple querystrings - a sorting querystring, and one for pagination.
As a solution, aforementioned dictionary is accessed again, any potential 'page' parameter is thrown out, and any other query string parameters that came along are reencoded in a query string and stored in a variable, which is passed to the template in the context.
Then that variable is used so that the pagination query string (when moving from one page to the next) is always appended to the correct url (either a url path with no query strings, or a url path + a query string for sorting)

To make the Html Select element show the option selected across page reloads, rather then reverting back to the default selection (even though the page has been reordered at this point according to another option), Javascript is used.
Before submitting any sorting query strings, the value of the selected sorting option is stored in a Session Storage variable.
On window load, the value of the Select HTML element is set to the variable stored in the Session Storage.