Added some WYSIWYG editor functionality to the admin pages as provided by CKeditor.
A few changes needed to be made: changing the body_text field on the Post model, as well as escaping the html in the tamplates ( using the safe filter in the post_detail template, and a combination of the safe and striptag filters in the home_page template.

I've also modified the 'cartesian product' by way of showing the new functionality.