Added a Contact form to the Contact Page
The only thing that happens, as it stands, when the form is submitted, is that a new entry in the db table is created, which is to be displayed in the Admin;
All fields will be displayed, along with the date when the entry was created (i.e. the form was submitted).
Redirection to the homepage ensues when the form is successfully submitted.