Implemented Ajax on the Home Page to update the body according to sorting and pagination, replacing the use of SessionStorage.
The home page was reloaded completely when a sorting option was selected or the next/previous page button was clicked.
SessionStorage was use to make the selected option persist on the UI across page reloads by first saving the selected option in local storage (for the duration of the session) before making the request for another page or for a re-sorted queryset,
and then restoring that on windowload.

It did the trick, but it was nowhere close to ideal.

With Ajax, a request is made asynchoronously, and the the server sends back an HTML string by rendering only the 'cards.html' template, but with the whole context of the 'homepage.html' accessible to it.
The problem is that since the homepage.html template doesn't get its context updated, only its html, the href attribute of the a element wrapped around the page button wouldn't be updated, and so each request for a new page would reset the queryset to the default sorting.

To get around that, the html for the article divs in the body isn't the only part replaced, but also the html for the pagination section, thus making it so that the url of the 'a' element always knows whether it's paginating a queryset with the default sorting,
or a query set that has been resorted, and can therefore preserve that (essentially ensuring the continuity of what was the case before when the home page would always get compltely reloaded with each request, rendered with an updated context).