Added a 'Categories' section on the homepage
There's a paragraph below the heading that's hidden and only appears when the heading is clicked;
A subsequent click will render the paragraph hidden again
The paragraph shows a list of those Categories that contain at least 1 article; the number of articles contained by each category is also shown in parentheses.
Each category is linked its own page.