Modified the cards.html template to use the include template tag
The cards.html template that's rendered and returned as html in response
to an Ajax request for a sorting or page-number change contained html
for both the pagination div and the 'cards' div below it in the body.

I split these into separate files : specifically, I took the
html for the pagination-div and put it into a 'pagination_snippet.html'
file, now included into the 'cards.html' file by means of an {%include%}
template tag.